backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
openai==1.77.0
pydantic==2.11.3
pydantic-settings==2.8.1
pypdf==5.4.0
python-docx==1.1.2
